+++ packages/atmosphere/src/shaders/sky.glsl
@@ -30,7 +30,8 @@ vec3 getSkyRadiance(
   const vec3 sunDirection,
   const vec3 moonDirection,
   const float moonAngularRadius,
-  const float lunarRadianceScale
+  const float lunarRadianceScale,
+  const float fragmentAngle
 ) {
   vec3 transmittance;
   vec3 radiance = GetSkyRadiance(
@@ -44,15 +45,9 @@ vec3 getSkyRadiance(
   // Rendering celestial objects without perspective doesn't make sense.
   #ifdef PERSPECTIVE_CAMERA
 
-  #if defined(SUN) || defined(MOON)
-  vec3 ddx = dFdx(rayDirection);
-  vec3 ddy = dFdy(rayDirection);
-  float fragmentAngle = length(ddx + ddy) / length(rayDirection);
-  #endif // defined(SUN) || defined(MOON)
-
   #ifdef SUN
   float viewDotSun = dot(rayDirection, sunDirection);
-  if (viewDotSun > cos(ATMOSPHERE.sun_angular_radius)) {
+  if (viewDotSun > cosSunAngularRadius) {
     float angle = acos(clamp(viewDotSun, -1.0, 1.0));
     float antialias = smoothstep(
       ATMOSPHERE.sun_angular_radius,