+++ packages/atmosphere/src/webgpu/ShadowLengthNode.ts
@@ -18,6 +18,15 @@
  * Modified from the original source code.
  */
 
+import {
+  DirectionalLight,
+  Matrix4,
+  PerspectiveCamera,
+  Vector2,
+  Vector3,
+  Vector4
+} from 'three'
+import type { CSMShadowNode } from 'three/examples/jsm/csm/CSMShadowNode.js'
 import { hash } from 'three/src/nodes/core/NodeUtils.js'
 import {
   float,
@@ -30,14 +39,9 @@ import {
   vec4
 } from 'three/tsl'
 import {
-  Matrix4,
   NodeMaterial,
   NodeUpdateType,
-  PerspectiveCamera,
   TempNode,
-  Vector2,
-  Vector3,
-  Vector4,
   type NodeBuilder,
   type NodeFrame,
   type TextureNode
@@ -45,11 +49,7 @@ import {
 import invariant from 'tiny-invariant'
 
 import { floorPowerOfTwo } from '@takram/three-geospatial'
-import {
-  OnBeforeFrameUpdate,
-  type CascadedShadowMapsNode,
-  type Node
-} from '@takram/three-geospatial/webgpu'
+import { OnBeforeFrameUpdate, type Node } from '@takram/three-geospatial/webgpu'
 
 import { getAtmosphereContext } from './AtmosphereContext'
 import { CoordinateNode } from './ShadowLengthNode/CoordinateNode'
@@ -69,7 +69,7 @@ export class ShadowLengthNode extends TempNode {
     return 'ShadowLengthNode'
   }
 
-  csmShadowNode: CascadedShadowMapsNode
+  csmShadowNode: CSMShadowNode
   viewZUnitNode!: TextureNode // Must be filterable
 
   sliceEndpointsNode: SliceEndpointsNode
@@ -101,10 +101,7 @@ export class ShadowLengthNode extends TempNode {
 
   private currentCascades = 0
 
-  constructor(
-    csmShadowNode: CascadedShadowMapsNode,
-    viewZUnitNode: TextureNode
-  ) {
+  constructor(csmShadowNode: CSMShadowNode, viewZUnitNode: TextureNode) {
     super('vec2')
     this.updateType = NodeUpdateType.FRAME // After CSM's updateBefore
 
@@ -133,6 +130,7 @@ export class ShadowLengthNode extends TempNode {
     if (camera == null || light == null) {
       return
     }
+    invariant(light instanceof DirectionalLight)
 
     const { lights } = csmShadowNode
     if (lights.length !== this.currentCascades) {