+++ packages/atmosphere/src/webgpu/ShadowLengthNode/EpipolarShadowLengthNode.ts
@@ -27,6 +27,7 @@ import {
   RGFormat,
   type PerspectiveCamera
 } from 'three'
+import type { CSMShadowNode } from 'three/examples/jsm/csm/CSMShadowNode.js'
 import {
   and,
   Break,
@@ -61,8 +62,7 @@ import {
   FnVar,
   Node,
   outputTexture,
-  raySpheresIntersections,
-  type CascadedShadowMapsNode
+  raySpheresIntersections
 } from '@takram/three-geospatial/webgpu'
 
 import { getAtmosphereContext } from '../AtmosphereContext'
@@ -101,7 +101,7 @@ export class EpipolarShadowLengthNode extends Node {
     return 'EpipolarShadowLengthNode'
   }
 
-  csmShadowNode!: CascadedShadowMapsNode
+  csmShadowNode!: CSMShadowNode
   coordinateNode!: TextureNode
   sliceUVDirectionNode!: TextureNode
   minMaxLevelsNode!: TextureNode
@@ -190,7 +190,7 @@ export class EpipolarShadowLengthNode extends Node {
 
     const sampleShadow = FnVar(
       (
-        shadowUVAndDepthInLightSpace: Node<'vec3'>,
+        shadowUVInLightSpace: Node<'vec2'>,
         cascadeIndex: Node<'int'>,
         depthInLightSpace: Node<'float'>
       ): Node<'float'> => {
@@ -199,7 +199,7 @@ export class EpipolarShadowLengthNode extends Node {
           If(cascadeIndex.equal(cascade), () => {
             isInLight.assign(
               shadowDepthNodes[cascade]
-                .sample(shadowUVAndDepthInLightSpace.xy)
+                .sample(shadowUVInLightSpace.xy)
                 .compare(depthInLightSpace)
             )
           })
@@ -213,7 +213,10 @@ export class EpipolarShadowLengthNode extends Node {
         cascadeIndex: Node<'int'>,
         rayEndCameraZ: Node<'float'>,
         cascadeStartCameraZ: Node<'float'>,
-        cascadeEndCameraZ: Node<'float'>
+        cascadeEndCameraZ: Node<'float'>,
+        fullRayLength: Node<'float'>,
+        viewDirection: Node<'vec3'>,
+        rayTopIntersection: Node<'vec2'>
       ): Node<'vec2'> => {
         const sliceIndex = uint(screenCoordinate.y)
         const minMaxShadowMapSize = int(minMaxLevelsNode.size().x).toConst()
@@ -247,8 +250,8 @@ export class EpipolarShadowLengthNode extends Node {
         const rayLength = distanceToRayEnd.sub(distanceToRayStart).toConst()
 
         const totalShadowLength = float(0).toVar()
+        const firstShadowMoment = float(0).toVar()
         const totalMarchedLength = float(0).toVar()
-        const distanceToFirstShadowedSection = float(-1).toVar()
 
         // WORKAROUND: We cannot use the early-return pattern.
         If(rayLength.lessThanEqual(10 * worldToUnit).not(), () => {
@@ -321,6 +324,9 @@ export class EpipolarShadowLengthNode extends Node {
           const shadowUVAndDepthStep = shadowTraceDirection
             .mul(shadowUVStepLength)
             .toConst()
+          const minMaxTextureYIndex = uint(sliceIndex)
+            .add(uint(cascadeIndex.sub(firstCascade)).mul(epipolarSliceCount))
+            .toConst()
 
           Loop(distanceMarchedInCascade.lessThan(rayLength), () => {
             // Clamp depth to a very small positive value to avoid z-fighting
@@ -385,9 +391,6 @@ export class EpipolarShadowLengthNode extends Node {
                 .toConst()
 
               // Load 1D min/max depths.
-              const minMaxTextureYIndex = uint(sliceIndex).add(
-                uint(cascadeIndex.sub(firstCascade)).mul(epipolarSliceCount)
-              )
               const minMaxTextureCoord = ivec2(
                 int(currentSamplePosition.shiftRight(currentTreeLevel)).add(
                   levelDataOffset
@@ -399,22 +402,15 @@ export class EpipolarShadowLengthNode extends Node {
                 .xy.toConst()
 
               // Determine if the ray section is fully lit or fully shadowed.
-              if (builder.renderer.reversedDepthBuffer) {
-                // With reversed depth buffer, the relations are reversed.
-                // maxDepth = closest to light
-                isInLight.assign(
-                  startEndDepthOnRaySection
-                    .greaterThanEqual(currentMinMaxDepth.yy)
-                    .all()
-                )
-              } else {
-                // minDepth = closest to light
-                isInLight.assign(
-                  startEndDepthOnRaySection
-                    .lessThanEqual(currentMinMaxDepth.xx)
-                    .all()
-                )
-              }
+              isInLight.assign(
+                builder.renderer.reversedDepthBuffer
+                  ? startEndDepthOnRaySection
+                      .greaterThanEqual(currentMinMaxDepth.yy)
+                      .all()
+                  : startEndDepthOnRaySection
+                      .lessThanEqual(currentMinMaxDepth.xx)
+                      .all()
+              )
               const isInShadow = (
                 builder.renderer.reversedDepthBuffer
                   ? startEndDepthOnRaySection
@@ -443,7 +439,7 @@ export class EpipolarShadowLengthNode extends Node {
             If(currentTreeLevel.lessThanEqual(minLevel), () => {
               isInLight.assign(
                 sampleShadow(
-                  currentShadowUVAndDepthInLightSpace,
+                  currentShadowUVAndDepthInLightSpace.xy,
                   cascadeIndex,
                   currentDepthInLightSpace
                 )
@@ -465,37 +461,24 @@ export class EpipolarShadowLengthNode extends Node {
             currentSamplePosition.addAssign(uint(1).shiftLeft(currentTreeLevel))
             distanceMarchedInCascade.addAssign(rayStepLengthUnit.mul(stepScale))
 
-            // Store the distance where the ray first enters the shadow.
-            distanceToFirstShadowedSection.assign(
-              distanceToFirstShadowedSection
-                .lessThan(0)
-                .and(isInLight.not())
-                .select(totalMarchedLength, distanceToFirstShadowedSection)
-            )
+            const shadowStepLength = integrationStep
+              .mul(isInLight.oneMinus())
+              .toConst()
+            const centerStepDistance = distanceToRayStart
+              .add(totalMarchedLength)
+              .add(integrationStep.mul(0.5))
 
-            totalShadowLength.addAssign(
-              integrationStep.mul(isInLight.oneMinus())
+            totalShadowLength.addAssign(shadowStepLength)
+            firstShadowMoment.addAssign(
+              shadowStepLength.mul(centerStepDistance)
             )
             totalMarchedLength.addAssign(integrationStep)
           })
         })
 
-        // If the whole ray is lit, set the distance to the first shadowed
-        // section to the total marched distance.
-        If(distanceToFirstShadowedSection.lessThan(0), () => {
-          distanceToFirstShadowedSection.assign(totalMarchedLength)
-        })
-
-        return vec2(
-          totalShadowLength,
-          distanceToFirstShadowedSection.add(distanceToRayStart)
-        )
+        return vec2(totalShadowLength, firstShadowMoment)
       }
     )
-
-    let fullRayLength: Node<'float'>
-    let viewDirection: Node<'vec3'>
-    let rayTopIntersection: Node<'vec2'>
 
     return Fn(() => {
       const { parametersNode } = getAtmosphereContext(builder)
@@ -506,6 +489,7 @@ export class EpipolarShadowLengthNode extends Node {
       const rayEndCameraZ = coordinate.z.toVar()
 
       const totalShadowLength = vec2(0).toVar()
+      const fullRayLength = float(0).toVar()
 
       // Skip samples with invalid screen coordinates.
       // WORKAROUND: We cannot use the early-return pattern.
@@ -524,8 +508,8 @@ export class EpipolarShadowLengthNode extends Node {
             camera
           ).toConst()
           const fullRay = rayTermination.sub(cameraPositionUnit).toConst()
-          fullRayLength = fullRay.length().toVar()
-          viewDirection = fullRay.div(fullRayLength).toConst()
+          fullRayLength.assign(fullRay.length())
+          const viewDirection = fullRay.div(fullRayLength).toConst()
 
           // Intersect the ray with the top of the atmosphere and the Earth:
           const intersections = getRaySphereIntersections(
@@ -534,7 +518,7 @@ export class EpipolarShadowLengthNode extends Node {
             vec3(0),
             vec2(topRadius, bottomRadius)
           ).toConst()
-          rayTopIntersection = intersections.xy
+          const rayTopIntersection = intersections.xy
           const rayBottomIntersection = intersections.zw
 
           // The camera is outside the atmosphere and the ray either does not
@@ -543,20 +527,20 @@ export class EpipolarShadowLengthNode extends Node {
           If(rayTopIntersection.y.greaterThan(0), () => {
             // Limit the ray length by the distance to the top of the
             // atmosphere if the ray does not hit terrain.
-            const originalRayLength = fullRayLength.toConst()
+            const rayLength = fullRayLength.toVar()
             If(rayEndCameraZ.greaterThanEqual(biasedCameraFar), () => {
-              fullRayLength.assign(HALF_FLOAT_MAX)
+              rayLength.assign(HALF_FLOAT_MAX)
             })
             // Limit the ray length by the distance to the point where the ray
             // exits the atmosphere.
-            fullRayLength.assign(min(fullRayLength, rayTopIntersection.y))
+            rayLength.assign(min(rayLength, rayTopIntersection.y))
             // If there is an intersection with the Earth surface, limit the
             // tracing distance to the intersection.
             If(rayBottomIntersection.x.greaterThan(0), () => {
-              fullRayLength.assign(min(fullRayLength, rayBottomIntersection.x))
+              rayLength.assign(min(rayLength, rayBottomIntersection.x))
             })
 
-            rayEndCameraZ.mulAssign(fullRayLength.div(originalRayLength))
+            rayEndCameraZ.mulAssign(rayLength.div(fullRayLength))
 
             Loop(
               {
@@ -574,32 +558,45 @@ export class EpipolarShadowLengthNode extends Node {
                   Break()
                 })
 
-                const shadowLength = processCascade(
-                  cascadeIndex,
-                  rayEndCameraZ,
-                  cascadeStartCameraZ,
-                  cascadeEndCameraZ
-                ).toConst()
-
-                // Keep shadowLength.y from the first cascade that has shadow.
-                totalShadowLength.y.assign(
-                  totalShadowLength.x
-                    .equal(0)
-                    .and(shadowLength.x.greaterThan(0))
-                    .select(shadowLength.y, totalShadowLength.y)
+                totalShadowLength.addAssign(
+                  processCascade(
+                    cascadeIndex,
+                    rayEndCameraZ,
+                    cascadeStartCameraZ,
+                    cascadeEndCameraZ,
+                    rayLength,
+                    viewDirection,
+                    rayTopIntersection
+                  )
                 )
-                totalShadowLength.x.addAssign(shadowLength.x)
               }
             )
-
-            // Set the full ray length to shadowLength.y to prevent
-            // interpolated values from being pulled towards 0.
-            If(totalShadowLength.x.equal(0), () => {
-              totalShadowLength.y.assign(fullRayLength)
-            })
           })
         }
       )
+
+      // totalShadowLength.y contains the first moment of shadow (m). It can be
+      // converted to the approximate distance to the first shadowed segment by:
+      //   y = m/x - 0.5x
+      //   m = \int_y^{y+x}{t}dt = yx + 0.5x^2
+      //   x = \int{f(t)}dt
+      // where x is totalShadowLength.x.
+      // This is unstable when x is small, but not a problem in practice because
+      // short shadow lengths affect less to the final output anyway.
+      If(totalShadowLength.x.greaterThan(1e-6), () => {
+        const distanceToFirstShadow = totalShadowLength.y
+          .div(totalShadowLength.x)
+          .sub(totalShadowLength.x.mul(0.5))
+          .clamp(0, fullRayLength.sub(totalShadowLength.x).max(0))
+          .toConst()
+        totalShadowLength.y.assign(
+          distanceToFirstShadow
+            .lessThan(worldToUnit)
+            .select(0, distanceToFirstShadow)
+        )
+      }).Else(() => {
+        totalShadowLength.y.assign(fullRayLength)
+      })
 
       return totalShadowLength
     })()